python_back/requirements.txt
flask==1.1.1
urllib3==1.25.7
mysql-connector==2.2.9
redis==3.3.11
Authlib>=0.13
sqlalchemy==1.3.12
psycopg2==2.8.4
sqlalchemy-migrate==0.13.0
passlib==1.7.2

ipython==7.11.0
ipdb==0.12.3